+++ test/hotspot/jtreg/gc/arguments/TestTargetSurvivorRatioFlag.java
@@ -254,7 +254,7 @@ public class TestTargetSurvivorRatioFlag {
         // Desired size of memory allocated at once
         public static final int CHUNK_SIZE = 1024;
         // Length of byte[] array that will have occupy CHUNK_SIZE bytes in heap
-        public static final int ARRAY_LENGTH = CHUNK_SIZE - Unsafe.ARRAY_BYTE_BASE_OFFSET;
+        public static final int ARRAY_LENGTH = CHUNK_SIZE - (int) Unsafe.ARRAY_BYTE_BASE_OFFSET;
 
         public static void main(String args[]) throws Exception {
             if (args.length != 1) {
